DL-2124 Removed old GG code and feature flag
