+++ conf/application.conf
@@ -149,5 +149,3 @@ auditing {
   }
 }
 
-# Feature toggle for EMACS: if set to false or absent then uses GG
-emacsFeatureToggle = true
