+++ project/FrontendBuild.scala
@@ -14,11 +14,11 @@ private object AppDependencies {
 
   val compile = Seq(
     ws,
-    "uk.gov.hmrc" %% "frontend-bootstrap" % "12.4.0",
+    "uk.gov.hmrc" %% "frontend-bootstrap" % "12.9.0",
     "uk.gov.hmrc" %% "play-partials" % "6.5.0",
     "uk.gov.hmrc" %% "domain" % "5.3.0",
-    "uk.gov.hmrc" %% "http-caching-client" % "8.1.0",
-    "uk.gov.hmrc" %% "auth-client" % "2.19.0-play-25"
+    "uk.gov.hmrc" %% "http-caching-client" % "8.3.0",
+    "uk.gov.hmrc" %% "auth-client" % "2.24.0-play-25"
   )
 
   trait TestDependencies {
@@ -33,7 +33,7 @@ private object AppDependencies {
         "org.pegdown" % "pegdown" % "1.6.0",
         "org.jsoup" % "jsoup" % "1.9.2" % scope,
         "com.typesafe.play" %% "play-test" % PlayVersion.current % scope,
-        "uk.gov.hmrc" %% "hmrctest" % "3.5.0-play-25",
+        "uk.gov.hmrc" %% "hmrctest" % "3.9.0-play-25",
         "org.mockito" % "mockito-all" % "1.10.19" % scope
       )
     }.test
